requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2
diffusers==0.18.3
accelerate==0.20.3
safetensors==0.3.1
opencv-python==4.7.0.72
pillow==9.5.0
numpy==1.24.3
huggingface-hub==0.16.4
ffmpeg-python==0.2.0
runpod==0.9.10
